reactor handling process of result

+++ mybatis-r2dbc/src/main/java/pro/chenggang/project/reactive/mybatis/support/r2dbc/executor/support/R2dbcStatementLog.java
@@ -80,7 +80,7 @@ public class R2dbcStatementLog {
      *
      * @param rows the rows
      */
-    public void logTotal(Integer rows) {
+    public void logTotal(Long rows) {
         debug("     Total: " + rows, false);
     }
